// src/features/tour/schemas/create-category.schema.ts
import { z } from "zod";

export const createCategorySchema = z.object({
  name: z.string().min(1, { message: "Tên không được để trống" }),
  status: z.string().min(1, { message: "Trạng thái không được để trống" }),
  description: z.string().optional().or(z.literal("")),
});